requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
coverage==6.4.2
Django==4.0.6
djangorestframework==3.13.1
Faker==13.15.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
